remove most UI changes from this PR
