+++ src/components/Form/index.js
@@ -6,6 +6,7 @@ import CrossPodWriteForm from './CrossPodWriteForm';
 import SetAclPermissionForm from './SetAclPermissionForm';
 import SetAclPermsDocContainerForm from './SetAclPermsDocContainerForm';
 import CheckAclPermsDocContainerForm from './CheckAclPermsDocContainerForm';
+import ManageUsers from './ManageUsers';
 
 /**
  * Components and functions related to Forms within project PASS
@@ -21,5 +22,6 @@ export {
   CrossPodWriteForm,
   SetAclPermissionForm,
   SetAclPermsDocContainerForm,
-  CheckAclPermsDocContainerForm
+  CheckAclPermsDocContainerForm,
+  ManageUsers
 };
