+++ src/contexts/index.js
@@ -1,7 +1,6 @@
 import SelectUserContext from './selectUserContext';
 import UserListContext, { UserListContextProvider } from './userListContext';
 import InboxMessageContext from './inboxMessageContext';
-import SignedInPodContext from './signedInPodContext';
 
 /**
  * The contexts module contains React Contexts that helps with user management
@@ -10,10 +9,4 @@ import SignedInPodContext from './signedInPodContext';
  * @namespace contexts
  */
 
-export {
-  SelectUserContext,
-  UserListContext,
-  InboxMessageContext,
-  SignedInPodContext,
-  UserListContextProvider
-};
+export { SelectUserContext, UserListContext, InboxMessageContext, UserListContextProvider };
