--- src/contexts/signedInPodContext.js
@@ -1,12 +1,0 @@
-import { createContext } from 'react';
-
-/**
- * React Context for the pod of the signed in User
- *
- * @name SelectUserContext
- * @memberof contexts
- */
-
-const SignedInPodContext = createContext('');
-
-export default SignedInPodContext;
