+++ src/utils/index.js
@@ -9,7 +9,7 @@
  */
 
 import docTypes from './frontend/form-helper';
-import { runNotification, clearProcessing } from './frontend/notification-helper';
+import runNotification from './frontend/notification-helper';
 import {
   uploadDocument,
   updateDocument,
@@ -30,7 +30,6 @@ import makeHandleFormSubmission from './frontend/FormSubmissionHelper';
 export {
   docTypes,
   runNotification,
-  clearProcessing,
   SOLID_IDENTITY_PROVIDER,
   uploadDocument,
   updateDocument,
